distbench-py/pyproject.toml
[project]
name = "distbench"
version = "0.1.0"
description = "Framework for distributed algorithms in Python"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Yigit" }
]
dependencies = [
    "pyyaml>=6.0",
    "msgpack>=1.0",
    "pynacl>=1.5.0",
    "click>=8.1",
]

[project.scripts]
distbench = "distbench.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.5",
    "ruff>=0.1",
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "types-PyYAML>=6.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
strict_concatenate = true

[[tool.mypy.overrides]]
module = "msgpack.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
